refactor: divide actions on separate files
