+++ src/constants/constants.ts
@@ -1,2 +1,3 @@
 export const PING_PONG_RATE = 2000;
 export const pasReg = new RegExp("(?=.*?[0-9]).{5,}", "g");
+export const THROTTLE_TIME = 1000;
