+++ tsconfig.json
@@ -1,8 +1,9 @@
 {
   "compileOnSave": false,
   "compilerOptions": {
-    "baseUrl": "src",
+    "rootDir": ".",
     "outDir": "dist",
+    "moduleResolution": "node",
     "forceConsistentCasingInFileNames": true,
     "strict": true,
     "noImplicitOverride": true,
@@ -14,15 +15,19 @@
     "downlevelIteration": true,
     "experimentalDecorators": true,
     "noEmit": true,
-    "allowImportingTsExtensions": true,
     "allowSyntheticDefaultImports": true,
-    "moduleResolution": "node",
     "importHelpers": true,
-    "target": "es6",
+    "target": "ES2020",
     "lib": [
       "ES2021.String"
-    ]
+    ],
   },
+  "include": [
+    "src/**/*.ts"
+  ],
+  "exclude": [
+    "node_modules",
+  ],
   "ts-node": {
     "esm": true,
   }
